+++ app/Http/Controllers/Admin/UserGuruController.php
@@ -3,13 +3,19 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Jurusan;
+use App\Models\JurusanTingkatKelas;
 use App\Models\Guru;
+use App\Models\Tingkat;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
 use Illuminate\Validation\Rule;
+use Yajra\DataTables\DataTables;
+use Illuminate\Support\Str;
 
 class UserGuruController extends Controller
 {
@@ -20,8 +26,8 @@ class UserGuruController extends Controller
      */
     public function index()
     {
-        $user = User::join('admin', 'users.username', '=', 'admin.username')
-                ->select('users.username', 'admin.*')
+        $user = User::join('guru', 'users.username', '=', 'guru.username')
+                ->select('users.username', 'guru.*')
                 ->where('users.id', Auth::user()->id)
                 ->first();
 
@@ -51,13 +57,22 @@ class UserGuruController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
+        $validator = Validator::make($request->all(), [
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'string', 'email', 'max:255', 'unique:guru'],
             'foto' => ['nullable', 'image', 'max:2048'],
         ]);
 
-        $prefix = '11'; // Prefix with 00
+        if($validator->fails()){
+            return response()->json([
+                'status' => false,
+                'modal_close' => false,
+                'message' => 'Data gagal ditambahkan. ' .$validator->errors()->first(),
+                'data' => $validator->errors()
+            ]);
+        }
+
+        $prefix = '11'; // Prefix with 11
         $uniqueId = str_pad(rand(0, 9999), 4, '0', STR_PAD_LEFT); // Generate a 4-digit unique ID
 
         $username = $prefix . $uniqueId;
@@ -70,14 +85,15 @@ class UserGuruController extends Controller
 
         $image_name = null; // Default value for image_name
 
-        if ($request->file('foto')) {
+        if ($request->hasFile('foto')) {
             $file = $request->file('foto');
             $extension = $file->getClientOriginalExtension();
-            $filename = 'guru-foto-' . $username . '.' . $extension;
+            $randomString = Str::random(3);
+            $filename = 'guru-foto-' . $username . '-' . $randomString . '.' . $extension;
             $image_name = $file->storeAs('file/img/guru', $filename, 'public');
         } else {
             $image_name = 'file/img/default/profile.png';
-        }
+        }        
 
         $hashedPassword = Hash::make($username);
 
@@ -94,8 +110,15 @@ class UserGuruController extends Controller
             'foto' => $image_name,
         ]);
 
-        return redirect('admin/input-guru')->with('success', 'User Berhasil Ditambahkan');
+        return response()->json([
+            'status' => true,
+            'modal_close' => false,
+            'message' => 'Data berhasil ditambahkan',
+            'data' => null
+        ]);
     }
+
+
 
     /**
      * Display the specified resource.
@@ -105,7 +128,19 @@ class UserGuruController extends Controller
      */
     public function show($id)
     {
-        //
+        $guru = Guru::find($id);
+
+        if ($guru) {
+            $user = User::where('username', $guru->username)->first();
+            if ($user) {
+                $guru->level_user = $user->level_user;
+            } else {
+                $guru->level_user = null;
+            }
+            return response()->json($guru);
+        } else {
+            return response()->json(['error' => 'Data not found'], 404);
+        }
     }
 
     /**
@@ -131,28 +166,42 @@ class UserGuruController extends Controller
     public function update(Request $request, $id)
     {
         // Validasi inputan
-        $validated = $request->validate([
+        $validator = Validator::make($request->all(), [
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'string', 'email', 'max:255', Rule::unique('guru')->ignore($id)],
-            'password' => ['nullable', 'string', 'min:4'],
             'foto' => ['nullable', 'image', 'max:2048'],
         ]);
 
-        // Cari data guru berdasarkan ID
+        if($validator->fails()){
+            return response()->json([
+                'status' => false,
+                'modal_close' => false,
+                'message' => 'Data gagal diubah. ' .$validator->errors()->first(),
+                'data' => $validator->errors()
+            ]);
+        }
+
+        // Cari data Guru berdasarkan ID
         $guru = Guru::find($id);
 
-        // Update data guru
-        $guru->name = $validated['name'];
-        $guru->email = $validated['email'];
+        if (!$guru) {
+            return response()->json(['error' => 'Guru tidak ditemukan.'], 404);
+        }
+
+        // Update data Guru
+        $guru->name = $request->input('name');
+        $guru->email = $request->input('email');
 
         // Proses upload dan update foto ke dalam server jika ada
         if ($request->hasFile('foto')) {
             $foto = $request->file('foto');
-            $fotoName = 'guru-foto-' . $guru->username . '.' . $foto->getClientOriginalExtension();
+            $extension = $foto->getClientOriginalExtension();
+            $randomString = Str::random(3); // Menghasilkan string acak sepanjang 3 karakter
+            $fotoName = 'guru-foto-' . $guru->username . '-' . $randomString . '.' . $extension;
 
             if ($guru->foto !== 'file/img/default/profile.png') {
                 Storage::disk('public')->delete($guru->foto);
-            }
+            }            
 
             // Simpan foto baru
             $foto->storeAs('file/img/guru', $fotoName, 'public');
@@ -168,7 +217,12 @@ class UserGuruController extends Controller
             $user->save();
         }
 
-        return redirect()->back()->with('success', 'Data guru berhasil diperbarui.');
+        return response()->json([
+            'status' => true,
+            'modal_close' => false,
+            'message' => 'Data berhasil diubah',
+            'data' => null
+        ]);
     }
 
     /**
@@ -180,20 +234,43 @@ class UserGuruController extends Controller
 
     public function destroy($id)
     {
-        $guru = Guru::find($id);
-        $user = User::where('username', $guru->username)->first();
+        try {
+            $guru = Guru::find($id);
+            $user = User::where('username', $guru->username)->first();
 
-        if ($guru->foto) {
-            Storage::disk('public')->delete($guru->foto);
-        }
-    
-        $guru->delete();
+            if ($guru->foto && $guru->foto !== 'file/img/default/profile.png') {
+                Storage::disk('public')->delete($guru->foto);
+            }
 
-        if ($user) {
-            $user->delete();
+            $guru->delete();
+
+            if ($user) {
+                $user->delete();
+            }
+
+            return response()->json([
+                'status' => true,
+                'modal_close' => false,
+                'message' => 'Data berhasil dihapus',
+                'data' => null
+            ]);
+        } catch (\Exception $e) {
+            return response()->json([
+                'status' => false,
+                'modal_close' => false,
+                'message' => 'Gagal menghapus data',
+                'data' => null
+            ]);
         }
+    }
     
-        return redirect()->back()->with('success', 'Data guru, user, dan file foto berhasil dihapus.');
+    public function data()
+    {
+        $data = Guru::selectRaw('id, username, name, email, foto');
+
+        return DataTables::of($data)
+                    ->addIndexColumn()
+                    ->make(true);
     }
     
 }
