+++ app/Http/Controllers/Admin/UserSiswaController.php
@@ -13,7 +13,10 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
 use Illuminate\Validation\Rule;
+use Yajra\DataTables\DataTables;
+use Illuminate\Support\Str;
 
 class UserSiswaController extends Controller
 {
@@ -24,13 +27,13 @@ class UserSiswaController extends Controller
      */
     public function index()
     {
-        $user = User::join('admin', 'users.username', '=', 'admin.username')
-                ->select('users.username', 'admin.*')
+        $user = User::join('siswa', 'users.username', '=', 'siswa.username')
+                ->select('users.username', 'siswa.*')
                 ->where('users.id', Auth::user()->id)
                 ->first();
 
         $siswa = User::join('siswa', 'users.username', '=', 'siswa.username')
-            ->where('users.level_user', 2)
+            ->where('users.level_user', 1)
             ->get();
 
         return view('admin.siswa', ['user' => $user])
@@ -55,13 +58,22 @@ class UserSiswaController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
+        $validator = Validator::make($request->all(), [
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'string', 'email', 'max:255', 'unique:siswa'],
             'foto' => ['nullable', 'image', 'max:2048'],
         ]);
 
-        $prefix = '22'; // Prefix with 00
+        if($validator->fails()){
+            return response()->json([
+                'status' => false,
+                'modal_close' => false,
+                'message' => 'Data gagal ditambahkan. ' .$validator->errors()->first(),
+                'data' => $validator->errors()
+            ]);
+        }
+
+        $prefix = '11'; // Prefix with 11
         $uniqueId = str_pad(rand(0, 9999), 4, '0', STR_PAD_LEFT); // Generate a 4-digit unique ID
 
         $username = $prefix . $uniqueId;
@@ -74,21 +86,22 @@ class UserSiswaController extends Controller
 
         $image_name = null; // Default value for image_name
 
-        if ($request->file('foto')) {
+        if ($request->hasFile('foto')) {
             $file = $request->file('foto');
             $extension = $file->getClientOriginalExtension();
-            $filename = 'siswa-foto-' . $username . '.' . $extension;
+            $randomString = Str::random(3);
+            $filename = 'siswa-foto-' . $username . '-' . $randomString . '.' . $extension;
             $image_name = $file->storeAs('file/img/siswa', $filename, 'public');
         } else {
             $image_name = 'file/img/default/profile.png';
-        }
+        }        
 
         $hashedPassword = Hash::make($username);
 
         User::create([
             'username' => $username,
             'password' => $hashedPassword,
-            'level_user' => 2,
+            'level_user' => 1,
         ]);
 
         Siswa::create([
@@ -98,8 +111,15 @@ class UserSiswaController extends Controller
             'foto' => $image_name,
         ]);
 
-        return redirect('admin/input-siswa')->with('success', 'User Berhasil Ditambahkan');
+        return response()->json([
+            'status' => true,
+            'modal_close' => false,
+            'message' => 'Data berhasil ditambahkan',
+            'data' => null
+        ]);
     }
+
+
 
     /**
      * Display the specified resource.
@@ -109,7 +129,19 @@ class UserSiswaController extends Controller
      */
     public function show($id)
     {
-        //
+        $siswa = Siswa::find($id);
+
+        if ($siswa) {
+            $user = User::where('username', $siswa->username)->first();
+            if ($user) {
+                $siswa->level_user = $user->level_user;
+            } else {
+                $siswa->level_user = null;
+            }
+            return response()->json($siswa);
+        } else {
+            return response()->json(['error' => 'Data not found'], 404);
+        }
     }
 
     /**
@@ -135,28 +167,42 @@ class UserSiswaController extends Controller
     public function update(Request $request, $id)
     {
         // Validasi inputan
-        $validated = $request->validate([
+        $validator = Validator::make($request->all(), [
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'string', 'email', 'max:255', Rule::unique('siswa')->ignore($id)],
-            'password' => ['nullable', 'string', 'min:4'],
             'foto' => ['nullable', 'image', 'max:2048'],
         ]);
 
-        // Cari data siswa berdasarkan ID
+        if($validator->fails()){
+            return response()->json([
+                'status' => false,
+                'modal_close' => false,
+                'message' => 'Data gagal diubah. ' .$validator->errors()->first(),
+                'data' => $validator->errors()
+            ]);
+        }
+
+        // Cari data Siswa berdasarkan ID
         $siswa = Siswa::find($id);
 
-        // Update data siswa
-        $siswa->name = $validated['name'];
-        $siswa->email = $validated['email'];
+        if (!$siswa) {
+            return response()->json(['error' => 'Siswa tidak ditemukan.'], 404);
+        }
+
+        // Update data Siswa
+        $siswa->name = $request->input('name');
+        $siswa->email = $request->input('email');
 
         // Proses upload dan update foto ke dalam server jika ada
         if ($request->hasFile('foto')) {
             $foto = $request->file('foto');
-            $fotoName = 'siswa-foto-' . $siswa->username . '.' . $foto->getClientOriginalExtension();
+            $extension = $foto->getClientOriginalExtension();
+            $randomString = Str::random(3); // Menghasilkan string acak sepanjang 3 karakter
+            $fotoName = 'siswa-foto-' . $siswa->username . '-' . $randomString . '.' . $extension;
 
             if ($siswa->foto !== 'file/img/default/profile.png') {
                 Storage::disk('public')->delete($siswa->foto);
-            }
+            }            
 
             // Simpan foto baru
             $foto->storeAs('file/img/siswa', $fotoName, 'public');
@@ -172,7 +218,12 @@ class UserSiswaController extends Controller
             $user->save();
         }
 
-        return redirect()->back()->with('success', 'Data siswa berhasil diperbarui.');
+        return response()->json([
+            'status' => true,
+            'modal_close' => false,
+            'message' => 'Data berhasil diubah',
+            'data' => null
+        ]);
     }
 
     /**
@@ -184,20 +235,43 @@ class UserSiswaController extends Controller
 
     public function destroy($id)
     {
-        $siswa = Siswa::find($id);
-        $user = User::where('username', $siswa->username)->first();
-
-        if ($siswa->foto) {
-            Storage::disk('public')->delete($siswa->foto);
-        }
+        try {
+            $siswa = Siswa::find($id);
+            $user = User::where('username', $siswa->username)->first();
     
-        $siswa->delete();
-
-        if ($user) {
-            $user->delete();
-        }
+            if ($siswa->foto && $siswa->foto !== 'file/img/default/profile.png') {
+                Storage::disk('public')->delete($siswa->foto);
+            }
     
-        return redirect()->back()->with('success', 'Data siswa, user, dan file foto berhasil dihapus.');
+            $siswa->delete();
+    
+            if ($user) {
+                $user->delete();
+            }
+    
+            return response()->json([
+                'status' => true,
+                'modal_close' => false,
+                'message' => 'Data berhasil dihapus',
+                'data' => null
+            ]);
+        } catch (\Exception $e) {
+            return response()->json([
+                'status' => false,
+                'modal_close' => false,
+                'message' => 'Gagal menghapus data',
+                'data' => null
+            ]);
+        }
+    }    
+    
+    public function data()
+    {
+        $data = Siswa::selectRaw('id, username, name, email, foto');
+
+        return DataTables::of($data)
+                    ->addIndexColumn()
+                    ->make(true);
     }
 
     public function kelasSiswa_index($t, $j)
